[poc/LEMS-3271-update-perseus-storybook] move components docs back to original location
