+++ .storybook/main.ts
@@ -40,6 +40,7 @@ const config: StorybookConfig = {
 
         // Docs for Perseus widgets, components, and renderers
         "../packages/perseus/src/__docs__/**/*.@(stories.ts|stories.tsx|mdx)",
+        "../packages/perseus/src/components/__docs__/**/*.@(stories.ts|stories.tsx|mdx)",
     ],
     addons: [
         "@storybook/addon-a11y",
