+++ packages/perseus/src/components/__docs__/button-group.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import ButtonGroup from "../../components/button-group";
+import ButtonGroup from "../button-group";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
