+++ packages/perseus/src/components/__docs__/fixed-to-responsive.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import {getDependencies} from "../../dependencies";
-import FixedToResponsive from "../../components/fixed-to-responsive";
+import FixedToResponsive from "../fixed-to-responsive";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
