+++ packages/perseus/src/components/__docs__/hud.stories.tsx
@@ -1,6 +1,6 @@
 import {action} from "storybook/actions";
 
-import Hud from "../../components/hud";
+import Hud from "../hud";
 
 import type {StoryObj, Meta} from "@storybook/react-vite";
