+++ packages/perseus/src/components/__docs__/lint.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import Lint from "../../components/lint";
+import Lint from "../lint";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
