+++ packages/perseus/src/components/__docs__/math-input.stories.tsx
@@ -1,6 +1,6 @@
 import {action} from "storybook/actions";
 
-import MathInput from "../../components/math-input";
+import MathInput from "../math-input";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
