+++ packages/perseus/src/components/__docs__/sortable.stories.tsx
@@ -1,4 +1,4 @@
-import Sortable from "../../components/sortable";
+import Sortable from "../sortable";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
