+++ packages/perseus/src/components/__docs__/tex.stories.tsx
@@ -1,4 +1,4 @@
-import TeX from "../../components/tex";
+import TeX from "../tex";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
