+++ packages/perseus/src/components/__docs__/zoomable-tex.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import ZoomableTex from "../../components/zoomable-tex";
+import ZoomableTex from "../zoomable-tex";
 
 import type {Meta, StoryObj} from "@storybook/react-vite";
